requirements.txt
dataclasses==0.6
Django==2.1.1
djangorestframework==3.8.2
djongo==1.2.29
pymongo==3.7.1
pytz==2018.5
sqlparse==0.2.4
